yaktın beni belongsTo (siparişler admine çekildi)

+++ app/Http/Controllers/Site/SiparislerController.php
@@ -24,29 +24,70 @@ class SiparislerController extends Controller
 
     public function add(Request $request){
 
-        foreach(Cart::content() as $productCartItem) {
-
-            $x = array(
-                "urun_id"       => $productCartItem->id,
-                "adet"          => $productCartItem->qty,
-                "fiyat"         => $productCartItem->total,
-                'image'         => $productCartItem->options->image,
-                "stok_kodu"     => $productCartItem->options->stok,
-                "kullanici_id"  => Auth::guard("uye")->id(),
-            );
-
-            SiparislerModel::create($x);
-        }
 
         $input = $request->all();
-
-        $input["siparis_id"] = SiparislerModel::all()->last()->id;
+        $input["orderNo"] = rand(100000000,999999999);
 
         OrderDetail::create($input);
 
-        Cart::destroy();
 
-        return  redirect()->route("site.index")->with("toast_success","Sipariş Başarılı Bir Şekilde Tamamlandı");
+        if ($input){
+
+            $sonid = OrderDetail::all()->last()->id;
+
+                foreach(Cart::content() as $productCartItem) {
+
+                    $x = array(
+                        "urunler_id"       => $productCartItem->id,
+                        "adet"          => $productCartItem->qty,
+                        "fiyat"         => $productCartItem->total,
+                        'image'         => $productCartItem->options->image,
+                        "stok_kodu"     => $productCartItem->options->stok,
+                        "kullanici_id"  => Auth::guard("uye")->id(),
+                        "siparisid"     => $sonid
+                    );
+
+                    SiparislerModel::create($x);
+                }
+
+                Cart::destroy();
+
+                return  redirect()->route("site.index")->with("toast_success","Sipariş Başarılı Bir Şekilde Tamamlandı");
+
+        }
+
+
+
+
+
+
+
+
+
+//        foreach(Cart::content() as $productCartItem) {
+//
+//            $x = array(
+//                "urun_id"       => $productCartItem->id,
+//                "adet"          => $productCartItem->qty,
+//                "fiyat"         => $productCartItem->total,
+//                'image'         => $productCartItem->options->image,
+//                "stok_kodu"     => $productCartItem->options->stok,
+//                "kullanici_id"  => Auth::guard("uye")->id(),
+//            );
+//
+//            SiparislerModel::create($x);
+//        }
+//
+//        $input = $request->all();
+//
+//        $input["siparisler_id"] = SiparislerModel::all()->last()->id;
+//        $input["orderNo"] = rand(100000000,999999999);
+//
+//        OrderDetail::create($input);
+//
+//        Cart::destroy();
+//
+//        return  redirect()->route("site.index")->with("toast_success","Sipariş Başarılı Bir Şekilde Tamamlandı");
 
     }
